requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
docopt==0.6.2
idna==3.3
influxdb==5.3.1
msgpack==1.0.3
pipreqs==0.4.11
pkg_resources==0.0.0
prometheus==0.3.0
protobuf-py3==2.5.1
psutil==5.8.0
python-dateutil==2.8.2
pytz==2021.3
quantile-python==1.1
requests==2.5.1
six==1.16.0
urllib3==1.26.7
yarg==0.1.9
python-dotenv==0.19.2
